+++ libft/submit/ft_memmove.c
@@ -12,7 +12,7 @@
 
 #include "libfh.h"
 
-void    *ft_memmove(void *dest, const void *src, unsigned int n)
+void    *ft_memmove(void *dest, const void *src, size_t n)
 {
     unsigned char       *d;
     const unsigned char *s;
