+++ libft/submit/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-unsigned int ft_strlen(const char *c)
+size_t  ft_strlen(const char *c)
 {
-    unsigned int i;
+    size_t  i;
 
     i = 0;
     while (c[i])
